requirements.txt
streamlit
pandas
Pillow
streamlit-image-coordinates
